Move the reference-input machinery next to the helper contract
A file of its own under lib/ was the wrong shelf twice over: lib/ is the request
pipeline — mapping resolution, request options, provider selection — and this is
called by providers, not by the pipeline; and providers had to reach for a
second module to implement one interface.

providers/providerHelper.ts already declares ImageTextToVideoTaskHelper, the
contract these providers implement, and already imports toArray, omit and the
error types this needs. Providers already import it at runtime for the base
class, so folding the machinery in costs no new import.

The task module would have read better still — imageTextToVideo.ts already owns
the caller-facing ReferenceInputs type — but providers import nothing from
tasks/ at runtime today, only types, and a value import would close a cycle
through getProviderHelper.

Pure move: fal's 23 call shapes and wavespeed's 6 are byte-identical.
